Fix reappearing fragment in SettingsActivity
Because SettingsActivity insists on doing setFragment every time it is
recreated, PreferenceFragment will call its root screen — therefore all
navigation history will be erased when SettingsActivity is recreated.

Additionally, this issue also causes fragments to ghost (appear on top
of another) if SettingsActivity is recreated in a subscreen.

This commit should address the issue by only setting the fragment if
there is no saved instance in the activity itself.