Don't give focus to an invisible search bar
When the search bar is hidden, the user can still tap on the topmost of
the launcher to trigger its focus event; this opens the soft keyboard
(if there is any), and will cause a breakdown of panel state.

To fix this, the commit toggles the focus state of the search bar. This
will cause the topmost area of the launcher to be non-interactive, but
this is better than having a ghost view triggering its events.