Don't animate searchContainer when in battery saver
Because animations are (for a lack of better term) ignored when in
battery saver mode, searchContainer never finishes its animation here
unless battery saver is toggled off. This causes the search bar to be
perpetually visible until it is manually dismissed.

To fix this, simply check for battery saver state. The check will
resolve to true in older APIs because they will always return false.