Allow keyboard to directly launch web search
When there is no app to launch, and search provider isn't set to the
indeterminate va,ue the 'search'/enter key in the keyboard will now act
as a shortcut to the web search button.

This commit also fixes the issue of the keyboard being dismissed when
pressing said key.