Implement a simple start screen for new users
Some new users are complaining about the confusing interface — some flat
out commenting that the launcher does not work because there was nothing
aside from the wallpaper to show.

The start screen (or dialogue) is a mean to eliminate the confusion by
showing a response that the launcher is working fine. The dialogue also
further aid the user in navigating, by prompting the default gestures
(tap and long-press).