Basic implementation of Work/foreign profile support
This commit refactors a huge part of the launcher to be able to load
apps from a different profile (e.g, Work profile).

Apps from different profile will be marked by a different scheme
(user-packageName), whereas same-profile apps are left to their old
scheme (packageName). This should hopefully preserve compatibility with
older version of the launcher as the old scheme is the only one used in
previous versions.