Filter items in whenTimerTicked
When filtering apps in whenChanged, there is a possibility that the
search text stays constant from it was before — this can cause the list
to be unfiltered. While this is the intended behaviour, users might want
to keep their previous search text, but retain the filter; this
behaviour was not plausible before.

This commit introduces a fix wherein the filter is performed every timer
tick, which makes the search operation synchronous with the visibility
of online search prompt as well as the 'delete all' button visibility
toggle. This may be perceived as slightly slower by some user, but it is
not a significant regression.

Fixes #48.