Use SAF for backup/restore in supported Android versions
Due to the requirement put by Android Q, we need to use SAF to utilise
backup/restore — even if this requirement has yet to be enforced.

Because SAF works all the way down to Android 4.4 (KK), we can also make
use of it for those versions. The legacy backup/restore screen will be
kept for older versions of Android.