Customisable desktop accent colour
Allow the user to set a custom accent colour for the desktop (used in
fast scroller and search prompt). This option is limited, however, and
it doesn't theme everything at the moment (especially those requiring
expensive Reflections).

Darker colours used in the fast scroller are automatically chosen
through ColorUtils.blendARGB(). This should ensure a decent-looking
colour mix, but more tweaking may be necessary in the future.

This commit also removes the theming from search bar cursor to prevent
it from looking mismatched when the user selects a different accent.