Handle long press event inside onActionStateChanged
Long press drag consumes the touch event meant for long press/click
(which is to be expected). However, this causes any action assigned to
long press to not be called - this isn't what we want.

To get around this, detect the long press inside of OnItemMoveListener,
precisely in onActionStateChanged. The hack used here is fairly simple;
we simply intercept the time taken by the event changes, and do a
guesstimate from there.

This fixes #25 and returns the previous behaviour past the regression
back.