Implement support for grid app list
After quite some time, finally a commit that satisfies #5; this commit
adds support for a grid app list layout. The grid size is currently not
customisable, but I plan on amending that in a future commit.

At the moment, both list and grid mode are available under the 'App
list' settings, which makes it sound confusing since it can also be NOT
a list. Let's leave that issue for last though.