Remove redundant keyboard control calls
Calling showSoftKeyboard during panel drag causes an inexplicable UX
issue wherein the keyboard appearance would cause the panel to lock up
in a perpetual anchoring and collapsing state (due to the keyboard
pushing the view up). This is a regression and it shouldn't happen.

Also further improve UX by not calling hideSoftKeyboard whenever the
panel is being slid around. It's janky that way.