Minor changes to soft keyboard handling
Use toggleSoftInput to show, and properly set a flag to hide the
keyboard. It seems like the compatibility is a bit icky when it comes to
these API, and that the current means used doesn't work with some
devices.

This is a blind guess, though several devices have reported a working
state after this commit is applied.