Better column scaling for grids
Grid shouldn't use the fixed 5 column width, as this may be too big or
too small depending on the display. To scale this properly, we have to
manually define the sizes for each screen resolution. I hate to do this,
but this is what may have to be done until a future commit addresses
this hackery.