[HACK] Move Widget-space to its own Page
Following the previous implementation for app list, widget-space is now
inline - user can swipe from the app list to access its contents.

This change comes with a few caveat, noticeably the buggy favourites
panel behaviour, as well as the slight regression in stability. This is
expected to be addressed in future releases.