Refactor PackageChangesReceiver mechanism
Having the receiver reload the entire launcher at all time whenever a
package is being installed/uninstalled will definitely be an annoyance,
not to mention it will be costly and expensive to do in a burst.

To handle this issue, this commit refactors the PackageChangesReceiver
to be a dumb broadcaster. Whenever a package change is received, it will
pass the news to any Pages subscribed to the IntentFilter.

This should once again be a workaround for the app
installation/uninstallation issue.