Correctly position PopupMenu in WebSearchProviderPreference
Passing view here will only give the reference to providerListView and
not the items. This causes the PopupMenu to appear beneath the list,
which may be problematic for longer lists.

To fix this, simply pass the child at the specified index. With this the
PopupMenu appears nicely beneath said child.