Check for AsyncTask status instead of synchronizing
Our issue here stems from the fact that FetchAppsTask has a tendency to
be fired twice: during onStart when it detects a new package and when
the receiver also detects a new package. This causes the app list to
double up with duplicate contents (which causes a crash with stable ID
enabled).

By synchronizing some issues were alleviated, but it was mostly meant to
bandaid the issue. Previous attempts to fix this by checking for
AsyncTask/FetchAppsTask status does exist as well, but it was botched
due to the repeated calls in onResume, onStart, and in the receiver.

This commit, alongside past commits attempting to resolve this issue,
should hopefully fix the issue described above.