Unbreak app pinning for foreign profiles
We shouldn't retrieve userPackageName by simply appending user serial
and package name itself, since this can cause an invalid package to be
retrieved.

This commit also addresses the force close that happens due to an
invalid package name being passed to getIcon() - before, the entirety of
the package name and the user serial was set as a parameter, which
getIcon() simply doesn't like.