Update app list without clearing it first
By calling clear() on the original list, we give it no chance to do a
minute, selective update. Having an app list update block the UI is a
bad taste of UX, and is generally annoying to look at. To fix this, we
updateDataSet() handle the new list, allowing it to animate the change
instead of defaulting to viewing the progress bar.