Fix theming in API 16 and lower
Due to the recent AppCompat update, API levels below 17 are having
issues with theming. Specifically, in setLocalNightMode, which has been
made inaccessible in API <17. To fix this, we do the long workaround of
using setTheme(), which has been proven to work without much hiccups.

Several XMLs have been restructured to reflect the separation, and to
increase clarity as to which styles/themes are used.