Reset searchContainer alpha value when in power saving mode
In the event that the user toggles power saving mode in their device
after hiding the panel (which should happen before power saving is
triggered), it is more than likely than the panel will stay invisible
(however, it can be interacted with). This is because the view is
visible, but the alpha value of searchContainer is still set to
transparent.

Fix this by resetting the alpha value back to 1f, which should allow it
to become visible again.