Improve LauncherAppWidgetHostView long press detection
Instead of using ACTION_UP, we can use ACTION_MOVE to detect the long
press. This has the advantage of automatically calling the long press
action when the duration is reached - thereby making the gesture more
natural (not needing the user to pull back/fire ACTION_UP).

* Also use ViewConfiguration duration. As it has been found, it is
possible to modify this duration through settings, so we should follow
said duration.