Make scrolling events work on widgets
Scrolling events were being consumed by the parent NestedScrollView.
This breaks the user interaction with widgets that can scroll (or have
the ability to be scrolled), as the scroll is instead passed to the
parent.

This hack prevents the NestedScrollView from intercepting the touch
events, which is then passed to the widget. The user may have difficulty
scrolling the list as whole, however. A future commit should address
that issue.