Allow theming and custom strings in ColorPreference
This is one giant flaming ball of mess and will definitely need to be
reviewed before it can get its flaming mess to production.

A future commit should address the strings situation. At the moment, we
aren't doing new strings to save up effort for the new major release
(and old strings are re-used instead).