Unbreak web search engine editing
The previous code assumes that when editing, the user is expected to
enter a new engine name. This is not realistic. The user might just
leave the name alone and edit only the URL.

Here, we make the assumption that:
* the URL doesn't need to be unique when adding a new engine
* when editing, the name OR the name AND URL can't be duplicates -
either of the two violates the unique property of the engine name

With this commit, editing should work again and new entries re-using a
duplicate name should be properly rejected.