Add new apps selectively
When the broadcast for a new app is received, the launcher will reload the entire app list. This is a waste of resources (and it's slow too), as the app information
can be retrieved without refreshing the list itself. This commit introduces a way to add the new app to the list without refreshing the entire list.