Don't add duplicate app when receiving update action
We don't have a check against existing apps when adding a new app. This
is problematic because Android fires REMOVED, ADDED, then CHANGED.
Effectively, the user may end up with duplicate entires of the app.

To fix this, we have a big check as well as a better update mechanism
that still conforms to the 'selective' nature of the addApp function.

* Also refactor PackageChangesReceiver to return usable actions