Actually guard against missing icon pack
When an icon pack is unexpectedly uninstalled while the user has not yet
set it back to 'default', the launcher will instead use some of the
leftover images from the old icon pack (most noticeably the back images
and the mask). This behaviour is unintended, as the icons should return
to default when a missing icon pack is detected.

To counter this (and unintentionally fix the logspam caused by the
missing package), we catch the error and re-set the icon pack to default
there, while also resetting some of the variables that may have been
modified by the missing icon pack.

* This commit also does a reformatting, but the reformatting should be
inconsequential to the actual purpose of the commit.