Use lifecycleScope from viewLifecycleOwner
Generally, this should be used when the fragment calls for coroutines in
its View lifecycle, so as to avoid missing context.

At the moment this commit is deployed to fix the crashing that occurs
due to fetchApps calling requireActivity when the fragment hasn't been
attached, which happens to a some testers in Play Store.