Fix the app menu once again
We can't use the simple list contains check here because it doesn't
account for the fact that the app can be in the list and the favourites
panel. This means that the app will have the same app menu no matter
where it is (which is not optimal since this would remove the
'Shorthand' and 'Hide' menu if the app has been pinned).

Thankfully, we can  just check if the menu was called from a pinned app
through the itemViewType property. By checking this, the menu returned
should properly reflect whether the app was actually pinned.