Separate 'open drawer' with 'open app list' for gestures
As per #87, we should separate the action of opening the drawer and app
list, seeing that they are both separate now. This commit should sort
the issue and add a separate entry for gesture actions.

Note that this commit also changes the default 'tap' action for new
users to 'open drawer' to retain old behaviour, but it won't change the
preferences that already persist.