Don't apply icon preferences in HiddenAppsPreference
By apply user preferences here, the icons can hide and actually change
depending on the icon pack. While this is great, there is absolutely no
need to do this (as it is possible to hide the icons themselves).

To fix this, separate the call needed for the main apps list and any
other list that may need loadApps(). This commit also simplifies the
call to loadApps(), which should hopefully make it more concise.