Prevent search text from bunching up/merging
If the user has a search query with space (for example: "hello world")
and they decide to append a character at the start of the query (i.e,
"!g hello world"), the text would instead get bunched together because
the space between the words was erased (the output will instead be "!g
helloworld").

This is primarily caused by the delete() being called. This call
prevents empty spaces from starting a query.

By checking for length and preventing the execution of s.delete() here,
the appended text won't disrupt the structure of the query, but it won't
change the behaviour we already have.