Don't check for panel scroll when calling showPinnedApps
We don't need to check for this because when the panel is set to static,
it won't hide due to the check in hidePinnedApps(). This check might
also cause a regression when the user first pins an app and the panel is
set to static, as showPinnedApps() won't be shown due to this check
returning false.