Early implementation of custom icon handling
As requested by #86, this feature allows the user to be able to select
an icon from icon packs, replacing the default and/or existing icons
provided by the current icon pack.

This implementation caches the icon into a PNG, saving it to the
launcher's files directory. This means that it can't be saved to when
backing up the launcher configuration,
but it will persist even if the source of this icon (the icon pack where
icon came from) is removed.

At the moment, there should be a consideration on the use of simple
index instead of having to cache an entire file to save on space, but
this works for now.